Added a test
The problem now is that the test leads .html files in the compiled
folder. I need to abstract the compiler parts into its own module or
trait, which configures the "compilation" directory. This would allow
compilation going to directories _outside_ of the app, which could be
useful if I decide to make it possible to recompile via a webhook.